recipe-ai/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langchain-ollama>=0.1.0
langchain-core>=0.2.0
python-dotenv>=1.0.0
